requirements.txt
fastapi
pillow
uvicorn

